refactor: DataPipeline, KakaoGeoService 코드 리팩토링
